questionnaire_service/questionnaire/views.py
from django.shortcuts import render
from rest_framework import generics
from .models import Question
from .serializers import QuestionSerializer
# Create your views here.


class QuestionListCreateAPIView(generics.ListCreateAPIView):
    queryset = Question.objects.all()
    serializer_class = QuestionSerializer

