questionnaire_service/requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.8.1
certifi==2023.5.7
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2022.1
Cython==3.0.6
distlib==0.3.6
dj-rest-auth==6.0.0
Django==5.1
django-allauth==0.63.2
django-crispy-forms==2.0
django-phonenumber-field==7.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
dnspython==2.5.0
drf-yasg==1.21.7
email-validator==2.1.0.post1
fastapi==0.109.1
filelock==3.12.0
graphene==3.2.2
graphene-django==3.1.3
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
openapi-codec==1.3.2
orjson==3.9.13
packaging==23.2
phonenumbers==8.13.15
Pillow==10.0.0
pipenv==2023.4.29
platformdirs==3.5.1
promise==2.3
psycopg2-binary==2.9.6
pydantic==2.6.0
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
python-dotenv==1.0.0
python-multipart==0.0.7
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.35.1
svgwrite==1.4.3
text-unidecode==1.3
Tree==0.2.4
typing_extensions==4.9.0
ujson==5.9.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.27.0.post1
uvloop==0.19.0
virtualenv==20.23.0
virtualenv-clone==0.5.7
watchfiles==0.21.0
websockets==12.0
whitenoise==6.5.0
